+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # nomnoml 0.1.0.9000
 
+- Add support for 'svg' output in 'nomnoml()'.
+
 - Fix issue introducing too much blank space in Rmd docs.
 
 # nomnoml 0.1.0